Codeforce 1238C
I ended up using a two states machine.
Since H may be too high, sequentially update each height is not a good
idea.

Since the lever can only change the current platform and the platform
below, and can't use magic on current platform, there are only two
states.
The first state: standing on current platform.
The second state: the platform right below current platform is also on.

In the second state, if the platform two levels below is also on,
the player can safely fall onto it and return to state one.
If the next on platform is further down, the player has to use a magic in
order to get to that platform. In this case, the state stays at two
since the player should be standing on the platform right above that
platform so there are still two consequtive platforms under the player
(including the one standing on).

Finally, add one magic use if the player is at state two and height is
higher than 2.