Codeforce 1245C, constant memory 'DP'
The answer is the production of every consecutive u or n segment's
number of possibility.
Segmenting consecutive u or n is easy, so the only work left is compute
number of possibility for a segment.
For a segment, the first char is 1, and it has 1 possible combination
with a trailing u or n.
For the second char, it will add "the number of trailing u or n" to
total possibility, and the number of possible combination that has
trailing u or n becomes the previous total combination.
For example:
----------------------------------
Char number  | 1 | 2 | 3 | 4 | 5 |
Num of Poss  | 1 | 2 | 3 | 5 | 8 |
Num of trail | 1 | 1 | 2 | 3 | 5 |
----------------------------------