Math, construction
There are two strategies:
1. Directly swap the two numbers without using the min element.
The numbers may need to be swapped multiple times. If the swapping chain
does not include the min element, then we don't need to check if both
numbers can be swapped with min element or not (i.e. one of the number
may be a multiplication of the min element).

2. Swap using the min element.
In this case, no matter how we re-order the swap chain, all the numbers
must be able to swap with min element.